sans: Add update-fonts.sh script
The current setup is a bit awkward, as the build system on its
own is not enough to build the project - it requires a source
archive that needs to up downloaded manually (or via CI).

Start centralizing this with a script that takes care of
downloading and extracting the sources.