build: Switch to meson build system
Now that all fonts are included in the repository, there isn't
much work for a build system. It is still useful though, as it
provides a standardized interface for which distributors etc
already have tooling.

However as most GNOME modules have replaced make with meson
by now, switch to that here as well.